Add Clear Linux support
This commit adds a new `Dockerfile` for the [Clear
Linux](https://clearlinux.org/) distribution.

I struggled with which version to use in this commit, because Clear
Linux is an
[auto-updating](https://clearlinux.org/documentation/clear-linux/concepts/swupd-about#updating)
distribution. As a result, I opted to use the `latest` Docker Hub tag.

The closest thing to a package manager that Clear Linux has is called
`swupd`, and the general idea is that _bundles_ of packages are
installed. I made sure that the bundles in this pull request install
each of the packages outlined in the `README.md`.